backend/requirements.txt
tensorflow==2.19.0
numpy==2.0.2
pandas==2.2.2
Pillow
scikit-learn
